requirements.txt
django==2.2.5
djangorestframework==3.10.3
drf_yasg==1.16.1
djangorestframework-jwt==1.11.0
rest-framework-generic-relations==1.2.2
psycopg2-binary==2.8.3
pycrypto==2.6.1
packaging==19.2